requirements.txt
Flask==2.2.3
discord.py==2.3.2
